Fix remove message, add spinner while loading
This commit fixes the remove, which was failing due to not pulling the JQ XHR variable out of the post request, so being unable to see the URL header. This also adds a new spinner which is inline along with the signin and task divs, and is hidden or shown in javascript. The signin button is moved to being hidden.

The spinner is convenient for the reload case - the auth token is read from the local storage, but a couple of API calls need to be made before the page is rendered. Previously it would just show the sign in button for this time, which can be for a half second up to a couple of seconds depending on the network connection. The spinner makes it clear that work is in progress, and is almost immediately hidden if there is no auth token in the local storage.